// src/utils/enums/permissions.enum.ts
export enum permissions {
    SUPER_ADMIN = 'Fenix-super_admin_access',
    ADMIN = 'Fenix-admin_access',
    PARAMETERIZER = 'Fenix-parametrizador_access',
    VALIDATOR = 'Fenix-validador_access',
    ANALYST = 'Fenix-analista_access',
    INVESTIGATOR = 'Fenix-investigador_access',
    // PATIENT = 'PACIENTE',
    // REPORTER = 'REPORTANTE',
    RISK_CONTROLLER = 'Fenix-controlador_riesgo_access',
    AUDITOR = 'Fenix-auditor_access',
    SP_PROGRAM = 'Fenix-programa_sp_access',
  }